requirements.txt
pytesseract==0.3.8
fuzzywuzzy==0.18.0
Flask_Mail==0.9.1
matplotlib==3.4.3
opencv-python-headless==4.5.3.56
numpy==1.21.2
Flask==1.0.2
Flask_PyMongo==2.3.0
pandas==1.3.2
python-dotenv==0.19.0
urllib3==1.4